Add bottom panel
It should be functional, but the current mobile styles drives it nuts.

It also drives me nuts.

It should appear immediately, but it doesn't do that in Chrome (desktop
and mobile). It looks ok on Firefox.

And now the masthead looks awkward, with the text right at the bottom of
the viewport on load.