Make bottom panel always visible on mobile
I was thinking of removing the annoying polygons on mobile, since they
tend to mess up somehow with the bottom panel. Hail WOW.js for allowing
the thing to be disabled on mobile! Now I prolly don't have to do that.

Since the panel is now always visible, there's no more need for its
accompanying script.