Fix polygon blocking the title
Setting the title to relative positioning moves it on top of the
big light blue polygon, solving half the problem.

The other problem is that the polygon has the same color as the title.
I initially intended to use an alpha background for the text, but it
looks awkward. So I just moved the polygon to the left on mobile.